renamed page name and updated dependencies
